fix mint test
